requirements.txt
mkdocs==1.6.1
mkdocs-material==9.5.35
mkdocstrings-python==1.11.1
pygame-ce==2.5.1
pytest==8.3.3
tomli==2.0.1